refactor: dom 조작 최소화를 위한 업데이트
