+++ src/view/containers/GameBox/gameBoxEvent.js
@@ -9,22 +9,7 @@ import GameModal from "../GameModal";
 import modalEvent from "../GameModal/modalEvent";
 
 const gameBoxEvent = () => {
-  const form = document.getElementById("money-submit");
 
-  form.onsubmit = function (event) {
-    event.preventDefault();
-    const money = event.target.money.value;
-    try {
-      LottoValidator.checkMoney(money);
-      store['lottos'] = generateLottos((money));
-      event.target.money.value = '';
-      Render.container("purchase-result", () => PurchaseResults(), () => { });
-      Render.container("input-winning-number", () => InputWinningNumberBox(), () => { inputWinningNumberEvent(); });
-      Render.container("game-result", () => GameModal(), () => modalEvent());
-    } catch (error) {
-      alert(error.message);
-    }
-  };
 };
 
 export default gameBoxEvent;
