--- src/view/containers/Header.js
@@ -1,8 +1,0 @@
-const Header = () => {
-  return `
-<div class="lotto-primary lotto-greyscale-1">
-  <div>🎱 행운의 로또</div>
-</div>  
-`;
-};
-export default Header;
